minor drawing changes
thickness of routes now increase based on the id of each vehicle, this is to allow us to visually separate lines when drawn between the same points.
to further aide visual separation, each vehicle is assigned a colour in which their routes will be drawn.

vehicles are also drawn in reverse order, this is so the thickest lines are drawn first, so no lines are obscured by another.